rename stop_state to stopping_state
